Fixes for LiveUpdate API after design meeting discussion (#1274)
* Design meeting fixes for LiveUpdate API
- Updated resource.store_resource to execute a callback after the resource has been stored to prepare for threaded storage of resources
- Updated documentation to reflect the API, temporarily disabled documentation for features we’re not ready to expose
- Updated LiveUpdate Settings dialog not to display the Defold Service setting since it’s not ready

* Fixed types so that buflen is now correctly using size_t instead of uint32_t

* Added checks for malloc failures when allocating data for digests and hex digests during store_resource.

* Fixed a bug where we would get BAD_ACCESS if a corrupt resource (that was smaller than the header) was sent to store_resource

+++ engine/liveupdate/src/liveupdate_private.h
@@ -24,7 +24,7 @@ namespace dmLiveUpdate
 
     uint32_t MissingResources(dmResource::Manifest* manifest, const dmhash_t urlHash, uint8_t* entries[], uint32_t entries_size);
 
-    void CreateResourceHash(dmLiveUpdateDDF::HashAlgorithm algorithm, const char* buf, uint32_t buflen, uint8_t* digest);
+    void CreateResourceHash(dmLiveUpdateDDF::HashAlgorithm algorithm, const char* buf, size_t buflen, uint8_t* digest);
 
 };
